fixing dark mode on reset password, moving components under cart to shopping-cart
