+++ src/app/_components/shopping-cart/CartIcon.js
